Deal with edge case where Graph API response is of type boolean
In very rare circumstances, the Graph API will return non-standard values, such as "false". When this happens, the JSON parser (correctly) reads this as a boolean.

As a result, when calling `response.get("error")`, an AttributeError is raised. This commit introduces an `InternalError` when Facebook isn't returning values consistent with the Graph Protocol.